requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-materializecss-form==1.1.14
django-registration-redux==2.6
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
